Make Nemotron nano recipe imports consistent with codebase (#1007)
* rm extra


* make nemotron nano v2 recipe imports consistent with codebase


* restore from bad rebase


* update


* add init


---------
